+++ lib/ardecy/harden/umask.rb
@@ -5,6 +5,7 @@ module Ardecy
     module Umask
       def self.exec(args)
         Umask::Profile.new(args).x
+        Umask::LoginDefs.new(args).x
         puts " ===> Corrected" if args[:fix]
       end
 
@@ -14,6 +15,16 @@ module Ardecy
           @file = '/etc/profile'
           @content = 'umask 027'
           @edit = 'umask'
+          @tab = 3
+        end
+      end
+
+      class LoginDefs < FileEdit
+        def initialize(args)
+          super
+          @file = '/etc/login.defs'
+          @content = 'UMASK 027'
+          @edit = 'UMASK'
           @tab = 3
         end
       end